Add Transaction Details view. Implement for XNV and copy to XMR. Change amounts to decimal to avoid scientific notation issue

+++ NervaOneWalletMiner/Objects/DataGrid/Account.cs
@@ -8,8 +8,8 @@ namespace NervaOneWalletMiner.Objects.DataGrid
         public string Label { get; set; } = string.Empty;
         public string AddressFull { get; set; } = string.Empty;
         public string AddressShort { get; set; } = string.Empty;        
-        public double BalanceLocked { get; set; }
-        public double BalanceUnlocked { get; set; }
+        public decimal BalanceLocked { get; set; }
+        public decimal BalanceUnlocked { get; set; }
         public Bitmap? WalletIcon { get; set; }
     }
 }
