+++ NervaOneWalletMiner/Rpc/Common/TransferDestination.cs
@@ -2,7 +2,7 @@
 {
     public class TransferDestination
     {
-        public double Amount { get; set; } = 0.0;
+        public decimal Amount { get; set; } = 0;
         public string Address { get; set; } = string.Empty;        
     }
 }
